requirements.txt
google-cloud
Flask==2.1.0
gunicorn==20.1.0
Werkzeug==2.3.7
appengine-python-standard>=1.0.0
